extract Databricks patch logic to separate module
